Use representer instead of recursive type converter
